refactor: [hw9 & hw10] folder structure according to onion arch
