+++ apps/weather-api/src/weather/domain-services/weather.interface.ts
@@ -1,0 +1,10 @@
+import { WeatherData } from '@weather/domain/weather.interface';
+
+export abstract class AbstractWeatherService {
+  abstract getWeather(city: string): Promise<WeatherData>;
+}
+
+export interface WeatherProvider {
+  name: string;
+  fetchWeatherData(city: string): Promise<WeatherData>;
+}
