+++ apps/weather-api/src/weather/domain/weather.interface.ts
@@ -3,32 +3,3 @@ export interface WeatherData {
   humidity: number;
   description: string;
 }
-
-export interface WeatherApiResponse {
-  current: {
-    temp_c: number;
-    humidity: number;
-    condition: {
-      text: string;
-    };
-  };
-}
-
-export interface OpenWeatherMapResponse {
-  main: {
-    temp: number;
-    humidity: number;
-  };
-  weather: Array<{
-    description: string;
-  }>;
-}
-
-export interface WeatherProvider {
-  name: string;
-  fetchWeatherData(city: string): Promise<WeatherData>;
-}
-
-export abstract class AbstractWeatherService {
-  abstract getWeather(city: string): Promise<WeatherData>;
-}
