+++ apps/weather-api/src/weather/infrastructure/weather-providers/weather-providers.interface.ts
@@ -1,0 +1,19 @@
+export interface WeatherApiResponse {
+  current: {
+    temp_c: number;
+    humidity: number;
+    condition: {
+      text: string;
+    };
+  };
+}
+
+export interface OpenWeatherMapResponse {
+  main: {
+    temp: number;
+    humidity: number;
+  };
+  weather: Array<{
+    description: string;
+  }>;
+}
