+++ apps/weather-api/test/mocks/weather-client.mock.ts
@@ -1,4 +1,4 @@
-import { WeatherData } from 'src/weather/domain/weather.interface';
+import { WeatherData } from '@weather/domain/weather.interface';
 
 export const mockWeatherClient = {
   fetchWeatherData: jest.fn().mockResolvedValue({
